move CharmState to base directory
