+++ paas_app_charmer/django/charm.py
@@ -48,7 +48,7 @@ class Charm(GunicornBase):  # pylint: disable=too-many-instance-attributes
         super().__init__(framework=framework, wsgi_framework="django")
         self.framework.observe(self.on.create_superuser_action, self._on_create_superuser_action)
 
-    def get_wsgi_config(self) -> BaseModel:
+    def get_framework_config(self) -> BaseModel:
         """Return Django framework related configurations.
 
         Returns:
